+++ packages/cache/src/Psl/Cache/Exception/ExceptionInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Psl\Cache\Exception;
+
+use Psl\Exception;
+
+/**
+ * Marker interface for all Cache component exceptions.
+ */
+interface ExceptionInterface extends Exception\ExceptionInterface {}